+++ package.json
@@ -13,6 +13,7 @@
     "bootstrap": "4.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "5.2",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
@@ -40,5 +41,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/react-router-dom": "^5.3.3"
   }
 }
