+++ src/utils/Menu.tsx
@@ -1,20 +1,22 @@
+import { NavLink } from "react-router-dom";
 
 export default function Menu(){
   const claseActiva = "active";
   return(
     <nav className="navbar navbar-expand-lg navbar-light bg-light">
       <div className="container-fluid">
-        <a className="navbar-brand" href="/">
+        <NavLink className="navbar-brand" activeClassName={claseActiva} 
+         to="/">
           React Peliculas
-        </a>
+        </NavLink>
         <div className="collapse navbar-collapse">
           <ul className="navbar-nav me-auto mb-2 mb-lg-0">
             <li className="nav-item">
-              <a
-                className="nav-link"
-                href="/generos"              >
+              <NavLink
+                className="nav-link" activeClassName={claseActiva} 
+                to="/generos">
                 Generos
-              </a>
+              </NavLink>
             </li>            
           </ul>
         </div>
